feat: extended application table, fix types original Event would not have questions property (which we need for table headers); luckily full event shape is used in all of the forms
