+++ frontend/src/org/components/EventForm/EventForm.tsx
@@ -1,5 +1,10 @@
 import { api } from 'app/services/bis'
-import type { Event, EventPayload, Location } from 'app/services/bisTypes'
+import type {
+  FullEvent,
+  Event,
+  EventPayload,
+  Location,
+} from 'app/services/bisTypes'
 import {
   EventPropagationImagePayload,
   Question,
@@ -565,7 +570,7 @@ export const EventForm: FC<{
       <Step name="přihlášky" hidden={!initialData?.id}>
         {initialData?.id && initialData?.name && (
           <ParticipantsStep
-            event={initialData as unknown as Event}
+            event={initialData as unknown as FullEvent}
             onlyApplications
           />
         )}
