+++ frontend/src/org/pages/EventApplications.tsx
@@ -1,10 +1,10 @@
-import { Event } from 'app/services/bisTypes'
+import { FullEvent } from 'app/services/bisTypes'
 import { Breadcrumbs } from 'components'
 import { ParticipantsStep } from 'org/components/EventForm/steps/ParticipantsStep'
 import { useOutletContext } from 'react-router-dom'
 
 export const EventApplications = () => {
-  const { event } = useOutletContext<{ event: Event }>()
+  const { event } = useOutletContext<{ event: FullEvent }>()
   return (
     <>
       <Breadcrumbs eventName={event.name} />
